long datatype no longer exists in Python 3.x, as int datatype will incorporate long. in Python 2.x, if long is supplied, int will automatically re-shape to a long. therefore, changing to int will work in 2 and 3.

+++ lib/ortho_functions.py
@@ -265,7 +265,7 @@ def buildParentArgumentParser():
     parser.add_argument("--no-pyramids", action='store_true', default=False, help='suppress calculation of output '
                                                                                   'image pyramids')
     parser.add_argument("--pyramid-type", choices=['near','cubic'], default='near', help='pyramid resampling strategy')
-    parser.add_argument("--ortho-height", type=long, help='constant elevation to use for orthorectification (value '
+    parser.add_argument("--ortho-height", type=int, help='constant elevation to use for orthorectification (value '
                                                           'should be in meters above the wgs84 ellipoid)')
